修改获取累计统计技术逻辑并且修改启动脚本文件

+++ src/main/java/com/tct/mapper/DeviceMapper.java
@@ -6,7 +6,7 @@ import java.util.List;
 import org.apache.ibatis.annotations.Param;
 
 public interface DeviceMapper {
-    int countByExample(DeviceExample example);
+    long countByExample(DeviceExample example);
 
     int deleteByExample(DeviceExample example);
